requirements.txt
python-telegram-bot==20.0
telethon==1.34.0
